Move BackupFinder in subfolder
